+++ src/utils.js
@@ -1,0 +1,5 @@
+const renderElements = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
+
+export {renderElements};
