+++ src/view/list.js
@@ -1,0 +1,5 @@
+// const createListTemplate = () => {return '<ul class="trip-events__list"></ul>';};
+
+const listTemplate = '<ul class="trip-events__list"></ul>';
+
+export {listTemplate};
