+++ webpack.config.js
@@ -1,17 +1,25 @@
 const path = require('path');
 const MAIN_PATH = './src/main.js';
 const OUTPUT_FILE_NAME = 'bundle.js';
-const directory =  path.resolve(__dirname, 'public');
+const publicDirectory =  path.resolve(__dirname, 'public');
+const viewDirectory = path.resolve(__dirname, 'src/view/');
+const utilsDirectory = path.resolve(__dirname, 'src/utils.js');
 
 module.exports = {
   entry: MAIN_PATH,
   devtool: 'source-map',
   output: {
     filename: OUTPUT_FILE_NAME,
-    path: directory,
+    path: publicDirectory,
   },  
   devServer: {
-    contentBase: directory,
+    contentBase: publicDirectory,
     watchContentBase: true,
-  }  
+  },
+  resolve: {
+    alias: {
+      view: viewDirectory,
+      utils: utilsDirectory,
+    },
+  }, 
 };
